Also create more realistic block and transaction ids and refactor
