+++ app/Providers/DevelopmentServiceProvider.php
@@ -6,9 +6,11 @@ namespace App\Providers;
 
 use Faker\Factory;
 use Faker\Generator;
+use Tests\FakerProviders\Wallet;
 use Illuminate\Support\Facades\App;
+use Tests\FakerProviders\Transaction;
 use Illuminate\Support\ServiceProvider;
-use Tests\Wallet;
+use Tests\FakerProviders\Block;
 
 final class DevelopmentServiceProvider extends ServiceProvider
 {
@@ -22,6 +24,8 @@ final class DevelopmentServiceProvider extends ServiceProvider
         if (App::environment(['local', 'testing']) === true) {
             $faker = Factory::create();
             $faker->addProvider(new Wallet($faker));
+            $faker->addProvider(new Transaction($faker));
+            $faker->addProvider(new Block($faker));
 
             $this->app->instance(Generator::class, $faker);
         }
