+++ tests/FakerProviders/Wallet.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests;
+namespace Tests\FakerProviders;
 
 use Faker\Generator;
 use Faker\Provider\Base;
